+++ docs/CHANGELOG.rst
@@ -7,6 +7,7 @@ Version 0.13.0dev (unreleased)
 * Add support for managing remote git repositories (Andreas Kaiser)
 * Add intersphinx to docs (Andreas Kaiser)
 * Add HTTP proxy support to speed up functional tests
+* Upgrade default Node.js version to 0.10.0
 
 Version 0.12.0 (2013-03-04)
 ---------------------------