+++ docs/CHANGELOG.rst
@@ -9,6 +9,7 @@ Version 0.13.0dev (unreleased)
 * Add HTTP proxy support to speed up functional tests
 * Upgrade default Node.js version to 0.10.0
 * Upgrade default Redis version to 2.6.11
+* Upgrade default Oracle JDK version to 7u17-b02
 
 Version 0.12.0 (2013-03-04)
 ---------------------------